Emit {{ .Release.Namespace }} for role subject
Active with --supress-namespace flag.
This is bit messy. Need to save state between recognising
kind: ServiceAccount and getting to the namespace,
so storing this on the context object.
Can only detect this condition while processing value output.
Relies on the fact that the keys are sorted as done in previous change.